switched to SSR
