+++ components/serverSections/music.jsx
@@ -1,0 +1,12 @@
+
+import { EPS } from '@/components/config';
+
+export default function Music() {
+
+    return (
+        <div>
+            <SectionTitle id="music" title="Our Music" />
+            <MusicSection music={EPS} />
+        </div>
+    )
+}
